rename kart_simulator en SingleplayerGame
