+++ client/output/__init__.py
@@ -1,5 +1,5 @@
 from .action_bar import BoxLayoutWithActionBar
-from .kart_simulator import SingleplayerGame
+from .SingleplayerGame import SingleplayerGame
 from .navigation_screen_manager import NavigationScreenManager, MyScreenManager
 from .outputFactory import OutputFactory
 from . import screens
